requirements.txt
attrs
lxml
packaging
pycairo
pygobject
